pyproject.toml
[project]
name = "cqhyxk"
version = "0.0.1"
description = "Python SDK for the 重庆环宇星科身份中台 V2.0"
readme = "README.md"
authors = [{name = "liudonghua123", email = "liudonghua123@gmail.com"}]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.7"
dependencies = [
    "requests>=2.25.0",
    "pydantic>=1.8.0,<2.0.0",
    "python-dotenv>=0.19.0",
]

[project.urls]
Homepage = "https://github.com/liudonghua123/cqhyxk"
Repository = "https://github.com/liudonghua123/cqhyxk"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"